+++ extensions/iceworks-lowcode-page-builder/package.json
@@ -97,6 +97,7 @@
     "@iceworks/material-service": "0.1.0-beta-2",
     "@iceworks/page-service": "0.1.0-beta-2",
     "@iceworks/material-ui": "0.1.0-beta-2",
+    "@iceworks/vscode-webview": "0.1.0-beta-2",
     "moment": "^2.24.0",
     "react": "^16.4.1",
     "react-dom": "^16.4.1",
