--- packages/block-add/src/__tests__/index.ts
@@ -1,6 +1,0 @@
-import { addBlock } from '..';
-
-test('generateProject', async () => {
-  const result = await addBlock(1);
-  expect(result).toBe(1);
-});
