--- packages/block-add/src/index.ts
@@ -1,5 +1,0 @@
-
-export async function addBlock(a: number): Promise<number> {
-  console.log('addBlock');
-  return a;
-}
