+++ packages/vscode-webview/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@iceworks/vscode-webview",
+  "version": "0.1.0-beta-2",
+  "files": [
+    "lib"
+  ],
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "devDependencies": {
+    "@types/vscode": "^1.45.0"
+  }
+}
