+++ packages/vscode-webview/src/index.ts
@@ -1,0 +1,2 @@
+export * from './webview';
+export * from './vscode';
