+++ packages/vscode-webview/src/template.html.ejs
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width,initial-scale=1,shrink-to-fit=no">
+  <meta name="theme-color" content="#000000">
+  <title>iceworks</title>
+  <link rel="stylesheet" type="text/css" href="${styleUri}">
+</head>
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="ice-container"></div>
+  <script nonce="${nonce}" src="${scriptUri}"></script>
+</body>
+</html>
