+++ packages/vscode-webview/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": "./",
     "rootDir": "src",
-    "outDir": "lib",
+    "outDir": "lib"
   },
   "exclude": ["src/__tests__", "lib"]
 }
